feat(no-confusing-non-null-assertion): disallow confusing non-null assertions
Non-null assertions `!` are disallowed when they immediately precede an equality comparison. This is
both unnecessary and confusing, since taken together the two can be confused for an inequality
comparison. This rule is somewhat redundant since non-null assertions are also banned outright, but
it is included for the benefit of projects that disable the no-non-null-assertion rule.