feat(no-implicit-any-catch): disallow use of `any` type in catch clause
BREAKING CHANGE: Use of the `any` type in a catch clause is no longer allowed. This is the default
type for the catch clause variable, and the only other permissible type is `unknown`, so catch
clause variables must be explicitly typed as `unknown`. The error can then be handled by narrowing
the type using type guards.