feat(no-loop-func): disallow function declarations with unsafe references inside loop statements
BREAKING CHANGE: Functions declared inside loop statements may not reference variables from the loop
scope declared with `var`, or variables declared with `let` or `var` from any parent scope where the
variable is modified inside or outside the loop. Such code is unlikely to behave as the author
intended and is usually indicative of an error.